pyproject.toml
[tool.poetry]
name = "installer-app"
version = "0.1.0"
description = ""
authors = ["jonathantoaf <yonitoaf123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.16.0"
pyyaml = "^6.0"
coloredlogs = "^15.0.1"
click = "^8.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
